# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Football News API",
    "description": "API for fetching football news from various sources with MySQL database scraping capabilities",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base path"
    }
  ],
  "paths": {
...
    "/news/90mins": {
      "get": {
        "summary": "Get news from 90mins",
        "description": "Returns football news articles from 90mins",
        "tags": ["📰 Basic News Endpoints (Original)"],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NinetyMinsArticle"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch 90mins news"
          }
        }
      }
    },
    "/news/onefootball": {
      "get": {
        "summary": "Get news from One Football",
        "description": "Returns football news articles from One Football",
        "tags": ["📰 Basic News Endpoints (Original)"],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OneFootballArticle"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch OneFootball news"
          }
        }
      }
    },
    "/news/espn": {
      "get": {
        "summary": "Get news from ESPN",
        "description": "Returns football news articles from ESPN",
        "tags": ["📰 Basic News Endpoints (Original)"],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ESPNArticle"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch ESPN news"
          }
        }
      }
    },
    "/news/goal": {
      "get": {
        "summary": "Get news from GOAL",
        "description": "Returns football news articles from GOAL",
        "tags": ["📰 Basic News Endpoints (Original)"],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GoalArticle"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch Goal.com news"
          }
        }
      }
    },
    "/news/fourfourtwo/epl": {
      "get": {
        "summary": "Get EPL news from FourFourTwo",
        "description": "Returns Premier League news articles from FourFourTwo",
        "tags": ["📰 Basic News Endpoints (Original)"],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FourFourTwoArticle"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch FourFourTwo EPL news"
          }
        }
      }
    },
    "/news/fourfourtwo/laliga": {
      "get": {
        "summary": "Get La Liga news from FourFourTwo",
        "description": "Returns La Liga news articles from FourFourTwo",
        "tags": ["📰 Basic News Endpoints (Original)"],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FourFourTwoArticle"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch FourFourTwo La Liga news"
          }
        }
      }
    },
    "/news/fourfourtwo/ucl": {
      "get": {
        "summary": "Get Champions League news from FourFourTwo",
        "description": "Returns Champions League news articles from FourFourTwo",
        "tags": ["📰 Basic News Endpoints (Original)"],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FourFourTwoArticle"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch FourFourTwo UCL news"
          }
        }
      }
    },
    "/news/fourfourtwo/bundesliga": {
      "get": {
        "summary": "Get Bundesliga news from FourFourTwo",
        "description": "Returns Bundesliga news articles from FourFourTwo",
        "tags": ["📰 Basic News Endpoints (Original)"],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FourFourTwoArticle"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch FourFourTwo Bundesliga news"
          }
        }
      }
    },
    "/scrape/{source}": {
      "post": {
        "summary": "🚀 Scrape full article content and save to MySQL",
        "description": "Scrapes full content from articles and saves to MySQL database. **Important:** Call the basic /news/{source} endpoint first to populate articles, then use this endpoint to scrape full content.",
        "tags": ["🔧 Enhanced Scraping Endpoints (New)"],
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "schema": { 
              "type": "string",
              "enum": ["90mins", "onefootball", "espn", "goal", "fourfourtwo-epl", "fourfourtwo-laliga", "fourfourtwo-ucl", "fourfourtwo-bundesliga"]
            },
            "description": "News source to scrape from"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "limit": {
                    "type": "number",
                    "default": 5,
                    "minimum": 1,
                    "maximum": 10,
                    "description": "Number of articles to scrape (max 10 recommended)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scraping completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScrapingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid source parameter"
          },
          "500": {
            "description": "Scraping failed"
          }
        }
      }
    },
    "/articles": {
      "get": {
        "summary": "📚 Get stored articles from MySQL database",
        "description": "Retrieve articles that have been scraped and stored in the MySQL database with pagination and filtering options",
        "tags": ["🔧 Enhanced Scraping Endpoints (New)"],
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "schema": { 
              "type": "string",
              "enum": ["90mins", "onefootball", "espn", "goal", "fourfourtwo-epl", "fourfourtwo-laliga", "fourfourtwo-ucl", "fourfourtwo-bundesliga"]
            },
            "description": "Filter by news source"
          },
          {
            "name": "category", 
            "in": "query",
            "schema": { 
              "type": "string",
              "enum": ["epl", "laliga", "ucl", "bundesliga", "general"]
            },
            "description": "Filter by category"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "number", "default": 20, "minimum": 1, "maximum": 100 },
            "description": "Number of articles per page"
          },
          {
            "name": "page",
            "in": "query", 
            "schema": { "type": "number", "default": 1, "minimum": 1 },
            "description": "Page number"
          }
        ],
        "responses": {
          "200": {
            "description": "Articles retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticlesResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database error"
          }
        }
      }
    },
    "/articles/{id}": {
      "get": {
        "summary": "📄 Get single article with full content",
        "description": "Retrieve a specific article by ID with complete content including full article text",
        "tags": ["🔧 Enhanced Scraping Endpoints (New)"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Article ID from MySQL database"
          }
        ],
        "responses": {
          "200": {
            "description": "Article found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FullArticle"
                }
              }
            }
          },
          "404": {
            "description": "Article not found"
          },
          "500": {
            "description": "Database error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NinetyMinsArticle": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Manchester United considering move for star striker"
          },
          "url": {
            "type": "string",
            "example": "https://www.90min.com/manchester-united-transfer-news"
          }
        }
      },
      "OneFootballArticle": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Liverpool win dramatic match against Arsenal"
          },
          "url": {
            "type": "string",
            "example": "https://onefootball.com/en/news/liverpool-win-dramatic-match-against-arsenal"
          },
          "img": {
            "type": "string",
            "example": "https://image.onefootball.com/liverpool-match.jpg"
          }
        }
      },
      "ESPNArticle": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Real Madrid secure Champions League victory"
          },
          "url": {
            "type": "string",
            "example": "https://www.espn.in/football/story/real-madrid-champions-league"
          },
          "img": {
            "type": "string",
            "example": "https://a.espncdn.com/photo/real-madrid-champions.jpg"
          }
        }
      },
      "GoalArticle": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://goal.com/lists/top-10-players-2023"
          },
          "modifiedTitle3": {
            "type": "string",
            "example": "Top 10 players of 2023"
          },
          "news_img": {
            "type": "string",
            "example": "https://images.goal.com/top-10-players.jpg"
          }
        }
      },
      "FourFourTwoArticle": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://www.fourfourtwo.com/premier-league/manchester-city-win-title"
          },
          "title": {
            "type": "string",
            "example": "Manchester City win Premier League title"
          },
          "news_img": {
            "type": "string",
            "example": "https://cdn.fourfourtwo.com/man-city-title.jpg"
          },
          "short_desc": {
            "type": "string",
            "example": "Manchester City have won their fourth consecutive Premier League title."
          }
        }
      },
      "ScrapingResponse": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string",
            "example": "90mins"
          },
          "totalFound": {
            "type": "number",
            "example": 10
          },
          "scraped": {
            "type": "number",
            "example": 5
          },
          "articles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FullArticle"
            }
          }
        }
      },
      "ArticlesResponse": {
        "type": "object",
        "properties": {
          "articles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StoredArticle"
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {"type": "number", "example": 1},
              "limit": {"type": "number", "example": 20},
              "total": {"type": "number", "example": 45},
              "totalPages": {"type": "number", "example": 3}
            }
          }
        }
      },
      "StoredArticle": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1,
            "description": "MySQL auto-increment ID"
          },
          "source": {
            "type": "string",
            "example": "90mins"
          },
          "title": {
            "type": "string",
            "example": "Manchester United considering move for star striker"
          },
          "original_title": {
            "type": "string",
            "example": "Man Utd eye striker move"
          },
          "url": {
            "type": "string",
            "example": "https://www.90min.com/manchester-united-transfer"
          },
          "image": {
            "type": "string",
            "example": "https://cdn.90min.com/image.jpg"
          },
          "description": {
            "type": "string",
            "example": "Manchester United are reportedly considering..."
          },
          "category": {
            "type": "string",
            "example": "general"
          },
          "scraped_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "word_count": {
            "type": "number",
            "example": 342
          },
          "is_active": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "FullArticle": {
        "allOf": [
          {"$ref": "#/components/schemas/StoredArticle"},
          {
            "type": "object",
            "properties": {
              "content": {
                "type": "string",
                "example": "Full article content goes here..."
              },
              "published_at": {
                "type": "string",
                "format": "date-time",
                "example": "2024-01-15T08:00:00.000Z"
              }
            }
          }
        ]
      },
...
    }
  }
}